# PCE replacement PoC for Ferma + OrientDB 