--- a/services/neo4j/persistence-tests/src/main/java/org/openkilda/persistence/ferma/OrientDbPersistenceManager.java
+++ b/services/neo4j/persistence-tests/src/main/java/org/openkilda/persistence/ferma/OrientDbPersistenceManager.java
@@ -19,10 +19,12 @@ import org.openkilda.persistence.PersistenceManager;
 import org.openkilda.persistence.TransactionManager;
 import org.openkilda.persistence.ferma.repositories.FermaRepositoryFactory;
 
+import com.google.common.annotations.VisibleForTesting;
 import com.orientechnologies.orient.core.db.ODatabaseType;
 import com.orientechnologies.orient.core.db.OrientDB;
 import com.syncleus.ferma.ext.orientdb.OrientTransactionFactory;
 import com.syncleus.ferma.ext.orientdb.impl.OrientTransactionFactoryImpl;
+import com.syncleus.ferma.tx.Tx;
 import org.apache.tinkerpop.gremlin.orientdb.OrientGraphFactory;
 
 /**
@@ -30,7 +32,8 @@ import org.apache.tinkerpop.gremlin.orientdb.OrientGraphFactory;
  */
 public class OrientDbPersistenceManager implements AutoCloseable {
     private final OrientDbConfig config;
-    private final OrientDB orientDb;
+    @VisibleForTesting
+    public final OrientDB orientDb;
 
     private transient volatile FermaTransactionManager transactionManager;
 
@@ -54,10 +57,10 @@ public class OrientDbPersistenceManager implements AutoCloseable {
                     OrientGraphFactory factory = new OrientGraphFactory(orientDb,
                             config.getDbName(),
                             ODatabaseType.valueOf(config.getDbType()),
-                            config.getUser(),
-                            config.getPassword());
+                            config.getDbUser(),
+                            config.getDbPassword());
                     OrientTransactionFactory txFactory =
-                            new OrientTransactionFactoryImpl(factory, true);
+                            new OrientTransactionFactoryImpl(factory, false);
                     transactionManager = new FermaTransactionManager(txFactory);
                 }
             }
@@ -71,6 +74,11 @@ public class OrientDbPersistenceManager implements AutoCloseable {
         if (transactionManager != null) {
             synchronized (this) {
                 if (transactionManager != null) {
+                    Tx tx = Tx.getActive();
+                    if (tx != null) {
+                        System.err.println("Closing an open TX on closing TX factory...");
+                        tx.close();
+                    }
                     ((OrientTransactionFactory) transactionManager.getTxFactory()).getFactory().close();
                     transactionManager = null;
                 }