--- a/services/neo4j/persistence-tests/src/main/resources/META-INF/services/org.neo4j.kernel.extension.KernelExtensionFactory
+++ b/services/neo4j/persistence-tests/src/main/resources/META-INF/services/org.neo4j.kernel.extension.KernelExtensionFactory
@@ -1,0 +1,1 @@
+org.openkilda.persistence.tests.neo4j.Neo4jUserProceduresExtension