The changes:
- Introduce the path computation benchmark tests. The topology is {3, 10, 30} * 30 = {90, 300, 900} nodes / vertices per an island, except crossIslandShortestPathBenchmark which has {3, 5, 10} * 30 = {90, 150, 300} vertices.

**Requirements for the solution:**
- Able to find N cheapest paths between 2 nodes / vertices in a graph. The cost is a function of relationship / edge properties and other provided data (e.g. diverse paths).
- Able to find a path with highest cost but cheaper than given.
- Has no issues with disconnected graphs.

**Potential solutions can be divided in 2 groups:**
- Graph traversal / computing is performed by DB (native support).
- A dedicated process / sub-system performs graph computing over data loaded from remote DB.

##### Neo4j with build-in traversal:
- Pure Cypher query doesn't fit the requirements: build-in functions “shortestPath” and “allShortestPaths” don't take into account the weight of relationships. We also had issues with disconnected graphs. 
- Graph Algorithms library / “algo.kShortestPaths” function: fast, support pre-loading of graphs. 
- The benchmark test with embedded Neo4j had a timeout issue with big graphs (Mesh with 900 nodes per island)

```
Benchmark                                                                           (maxHops)  (topologySize)  Mode  Cnt    Score     Error  Units
Neo4jNativeCheapestPathBenchmark.cheapestPathOnCircleTopologyBenchmark                     50               3    ss   10   17.684 ±   8.231  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnCircleTopologyBenchmark                     50              10    ss   10   17.783 ±   3.103  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnCircleTopologyBenchmark                     50              30    ss   10   35.202 ±  46.220  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnCircleTopologyBenchmark                    300               3    ss   10   11.768 ±   3.704  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnCircleTopologyBenchmark                    300              10    ss   10   27.332 ±  10.449  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnCircleTopologyBenchmark                    300              30    ss   10   90.578 ±  19.567  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnCircleTopologyBenchmark                   1000               3    ss   10   14.142 ±   4.411  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnCircleTopologyBenchmark                   1000              10    ss   10   27.374 ±  10.383  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnCircleTopologyBenchmark                   1000              30    ss   10   96.788 ±  21.846  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnMeshTopologyBenchmark                       50               3    ss   10   29.447 ±   3.698  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnMeshTopologyBenchmark                       50              10    ss   10   63.908 ±   3.846  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnMeshTopologyBenchmark                      300               3    ss   10   32.544 ±   5.386  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnMeshTopologyBenchmark                      300              10    ss   10   92.573 ±  12.767  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnMeshTopologyBenchmark                     1000               3    ss   10   31.607 ±   3.771  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnMeshTopologyBenchmark                     1000              10    ss   10   93.174 ±  16.992  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                       50               3    ss   10    9.779 ±   0.879  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                       50              10    ss   10   19.677 ±   7.154  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                       50              30    ss   10   26.928 ±   6.026  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                      300               3    ss   10   11.031 ±   2.667  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                      300              10    ss   10   17.801 ±   6.548  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                      300              30    ss   10   27.465 ±   5.936  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                     1000               3    ss   10   11.120 ±   1.501  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                     1000              10    ss   10   17.807 ±   2.156  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                     1000              30    ss   10   29.048 ±   7.102  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                       50               3    ss   10   10.412 ±   2.206  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                       50              10    ss   10   11.447 ±   2.406  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                       50              30    ss   10   16.797 ±   6.201  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                      300               3    ss   10   10.781 ±   1.681  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                      300              10    ss   10   11.733 ±   3.285  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                      300              30    ss   10   14.623 ±   2.750  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                     1000               3    ss   10   11.461 ±   3.923  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                     1000              10    ss   10   12.470 ±   6.140  ms/op
Neo4jNativeCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                     1000              30    ss   10   15.362 ±   4.419  ms/op
Neo4jNativeCheapestPathBenchmark.concurrent30CheapestPathOnCircleTopologyBenchmark        300              10    ss   10  194.081 ±  98.947  ms/op
Neo4jNativeCheapestPathBenchmark.concurrent50CheapestPathOnCircleTopologyBenchmark        300              10    ss   10  279.029 ± 137.900  ms/op
Neo4jNativeCheapestPathBenchmark.concurrent5CheapestPathOnCircleTopologyBenchmark         300              10    ss   10   37.823 ±   5.161  ms/op
Neo4jNativeCheapestPathBenchmark.crossIslandCheapestPathBenchmark                          50               3    ss   10   11.807 ±   4.287  ms/op
Neo4jNativeCheapestPathBenchmark.crossIslandCheapestPathBenchmark                          50               5    ss   10   12.246 ±   1.289  ms/op
Neo4jNativeCheapestPathBenchmark.crossIslandCheapestPathBenchmark                          50              10    ss   10   19.093 ±   8.717  ms/op
Neo4jNativeCheapestPathBenchmark.crossIslandCheapestPathBenchmark                         150               3    ss   10   11.554 ±   1.689  ms/op
Neo4jNativeCheapestPathBenchmark.crossIslandCheapestPathBenchmark                         150               5    ss   10   12.903 ±   2.649  ms/op
Neo4jNativeCheapestPathBenchmark.crossIslandCheapestPathBenchmark                         150              10    ss   10   16.420 ±   0.727  ms/op
Neo4jNativeCheapestPathBenchmark.crossIslandCheapestPathBenchmark                         300               3    ss   10   12.091 ±   3.178  ms/op
Neo4jNativeCheapestPathBenchmark.crossIslandCheapestPathBenchmark                         300               5    ss   10   12.534 ±   1.761  ms/op
Neo4jNativeCheapestPathBenchmark.crossIslandCheapestPathBenchmark                         300              10    ss   10   16.390 ±   1.357  ms/op
```

- It's possible to combine the fast build-in functions for traversals and Gremlin (via BOLT driver) for Kilda persistence layer.

##### OrientDB with build-in traversal:
- Orient SQL doesn't work for us: build-in functions “dijkstra” and “astar” work only with stored fields as the edge weight, so no way to modify the cost per request needs (e.g. diverse paths).

##### Current PCE (with Neo4j as the back-end): 
- Can be customized for any our needs. Slow, as loads the graph from DB into memory on every traversal.
- Can be scaled out separately from back-end storage.
- The benchmark test with embedded Neo4j had a timeout issue with big graphs (Mesh with 900 nodes per island)

```
Benchmark                                                                       (maxHops)  (topologySize)  Mode  Cnt     Score     Error  Units
KildaPceCheapestPathBenchmark.cheapestPathOnCircleTopologyBenchmark                    50               3    ss   10   209.980 ± 103.911  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnCircleTopologyBenchmark                    50              10    ss   10   344.666 ± 188.589  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnCircleTopologyBenchmark                   300               3    ss   10   172.915 ±  91.564  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnCircleTopologyBenchmark                   300              10    ss   10   373.492 ± 209.345  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnCircleTopologyBenchmark                  1000               3    ss   10   165.273 ±  87.663  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnCircleTopologyBenchmark                  1000              10    ss   10   423.587 ± 214.080  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnMeshTopologyBenchmark                      50               3    ss   10   781.480 ±  210.294  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnMeshTopologyBenchmark                      50              10    ss   10  2415.245 ±  370.347  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnMeshTopologyBenchmark                     300               3    ss   10   995.472 ±  627.050  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnMeshTopologyBenchmark                     300              10    ss   10  2549.941 ±  236.647  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnMeshTopologyBenchmark                    1000               3    ss   10   809.043 ±  214.469  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnMeshTopologyBenchmark                    1000              10    ss   10  2546.059 ±  675.539  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                      50               3    ss   10   170.763 ±  75.875  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                      50              10    ss   10   351.142 ± 196.579  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                      50              30    ss   10   724.036 ± 292.384  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                     300               3    ss   10   147.250 ±  87.706  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                     300              10    ss   10   333.530 ± 232.389  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                     300              30    ss   10   706.535 ± 280.388  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                    1000               3    ss   10   160.197 ±  85.429  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                    1000              10    ss   10   338.409 ± 189.454  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                    1000              30    ss   10   886.722 ± 727.096  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                      50               3    ss   10    70.785 ±  38.306  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                      50              10    ss   10   143.500 ±  79.910  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                      50              30    ss   10   275.282 ± 139.465  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                     300               3    ss   10    83.189 ±  37.369  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                     300              10    ss   10   140.892 ±  85.895  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                     300              30    ss   10   288.696 ± 121.164  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                    1000               3    ss   10    81.016 ±  55.993  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                    1000              10    ss   10   134.126 ±  78.706  ms/op
KildaPceCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                    1000              30    ss   10   296.937 ± 231.109  ms/op
KildaPceCheapestPathBenchmark.concurrent30CheapestPathOnCircleTopologyBenchmark       300              10    ss   10  4583.173 ± 1565.971  ms/op
KildaPceCheapestPathBenchmark.concurrent50CheapestPathOnCircleTopologyBenchmark       300              10    ss   10  7156.124 ± 2148.781  ms/op
KildaPceCheapestPathBenchmark.concurrent5CheapestPathOnCircleTopologyBenchmark        300              10    ss   10  1089.656 ± 642.178  ms/op
KildaPceCheapestPathBenchmark.crossIslandCheapestPathBenchmark                         50               3    ss   10   199.484 ±  97.079  ms/op
KildaPceCheapestPathBenchmark.crossIslandCheapestPathBenchmark                         50               5    ss   10   225.550 ± 129.724  ms/op
KildaPceCheapestPathBenchmark.crossIslandCheapestPathBenchmark                         50              10    ss   10   410.153 ± 222.694  ms/op
KildaPceCheapestPathBenchmark.crossIslandCheapestPathBenchmark                        150               3    ss   10   190.095 ± 124.808  ms/op
KildaPceCheapestPathBenchmark.crossIslandCheapestPathBenchmark                        150               5    ss   10   267.875 ± 185.341  ms/op
KildaPceCheapestPathBenchmark.crossIslandCheapestPathBenchmark                        150              10    ss   10   336.283 ± 165.097  ms/op
KildaPceCheapestPathBenchmark.crossIslandCheapestPathBenchmark                        300               3    ss   10   267.345 ± 132.707  ms/op
KildaPceCheapestPathBenchmark.crossIslandCheapestPathBenchmark                        300               5    ss   10   316.763 ± 226.933  ms/op
KildaPceCheapestPathBenchmark.crossIslandCheapestPathBenchmark                        300              10    ss   10   451.353 ± 268.925  ms/op
```

##### Gremlin Server with embedded Neo4j:
- Requires multi-instance deployment to combine Gremlin traversals and transactional operations via  BOLT / HTTP connectors (an approach with Neo4j HA configuration).

##### Neo4j with neo4j-gremlin-bolt:
- Use Thinkerpop API with a remote Neo4j server via BOLT protocol. Both for Gremlin traversals and transactional operations.
- The benchmark test with embedded Neo4j had a timeout issue with big graphs (Mesh with 900 nodes per island)

```
Benchmark                                                                              (maxHops)  (topologySize)  Mode  Cnt     Score      Error  Units
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnCircleTopologyBenchmark                     50               3    ss   10    69.824 ±   32.274  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnCircleTopologyBenchmark                     50              10    ss   10   140.524 ±   52.475  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnCircleTopologyBenchmark                    300               3    ss   10    71.003 ±   24.123  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnCircleTopologyBenchmark                    300              10    ss   10   213.140 ±   43.419  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnCircleTopologyBenchmark                   1000               3    ss   10    70.706 ±   23.926  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnCircleTopologyBenchmark                   1000              10    ss   10   205.586 ±   65.842  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnMeshTopologyBenchmark                       50               3    ss   10   985.028 ±   88.382  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnMeshTopologyBenchmark                      300               3    ss   10  1020.334 ±   67.821  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnMeshTopologyBenchmark                     1000               3    ss   10   974.818 ±   86.401  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                       50               3    ss   10    56.363 ±   22.793  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                       50              10    ss   10   166.046 ±   78.921  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                       50              30    ss   10   591.279 ±  382.852  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                      300               3    ss   10    63.658 ±   15.271  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                      300              10    ss   10   179.815 ±   67.334  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                      300              30    ss   10   554.781 ±  256.173  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                     1000               3    ss   10    61.824 ±   27.493  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                     1000              10    ss   10   168.012 ±   44.697  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnStarTopologyBenchmark                     1000              30    ss   10   564.689 ±  181.490  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                       50               3    ss   10    25.879 ±    4.163  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                       50              10    ss   10    45.222 ±    8.900  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                       50              30    ss   10    98.189 ±   40.904  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                      300               3    ss   10    24.982 ±    4.747  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                      300              10    ss   10    52.544 ±   16.233  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                      300              30    ss   10   109.719 ±   51.606  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                     1000               3    ss   10    26.047 ±    5.993  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                     1000              10    ss   10    48.986 ±   13.272  ms/op
Neo4jTinkerpopCheapestPathBenchmark.cheapestPathOnTreeTopologyBenchmark                     1000              30    ss   10   114.673 ±   40.035  ms/op
Neo4jTinkerpopCheapestPathBenchmark.concurrent30CheapestPathOnCircleTopologyBenchmark        300              10    ss   10  2700.741 ± 1641.735  ms/op
…