@sergii-iakovenko 2019-09-05
![shortest-path-tests-topologies](https://user-images.githubusercontent.com/7511304/64341408-f4f04b00-cff0-11e9-9359-09aeee4f24e7.png)